+++ packages/api/internal/api/spec.gen.go
@@ -19,7 +19,7 @@ import (
 var swaggerSpec = []string{
 
 	"H4sIAAAAAAAC/+xbbW/bOPL/KgT//5fa2M12F3t+l7S9vWDbbtCkewcEwYGWxjE3FKklKSdG4O9+4IMk",
-	"SqJs2XlAgrtXdcXhcDjzm+FwhnnAqcgLwYFrhWcPuCCS5KBB2v/NS8qys4/mJ+V4hguilzjBnOSAZ/Vo",
+	"SqJs2XlAcrhXdcXhcDjzm+FwhnnAqcgLwYFrhWcPuCCS5KBB2v/NS8qys4/mJ+V4hguilzjBnOSAZ/Vo",
 	"giX8VVIJGZ5pWUKCVbqEnJhpel0YUqUl5Td4s0mwIjybi/tBrs34fnw15AUjGgYZBwT7cN4YYlUIrsDq",
 	"5P10av5JBdfAtflJioLRlGgq+ORPJbj51vD7fwkLPMP/N2kUPXGjavJJSiHdGhmoVNLCMMEzfEoyZEQE",
 	"pfEmwe+n755/zZNSL4FrzxWBozOLv3/+xb8KjRai5JlZ8aeXUPEFyBXIapubCgLWxh/Ov38QpVu6PevD",
@@ -46,21 +46,21 @@ var swaggerSpec = []string{
 	"VdCViEG/ZjsxRM5EE3+vcWtFxf5MlUaEsboo0EyJbOIiGIztY3T/oI7e49KqToVyc92/Iva7DbXm2BpJ",
 	"0KXkkA1s1ap2Oka10/3MUHdWttMaotCnrEJC7F9dm4RSE3NGXuFGcKOIQqhY28Qe9YjUJRNbTOkE2LaB",
 	"z4XqWNji8FRk6ydrDgU1/k076PkcuQOrp+vCtZbtZHCdSqdPkwLnY+u3jJFWKJg81HWwjYMNAx05mn+j",
-	"jDXw6YHlo51Ww+UiqK2FHe0Bx25IJk1VzojfMf/7SLrdMdYtZSxqq/H69/3WXbTvX8afo3H6V9BbjBGG",
-	"5meyxNP1h7vB/L82dA+65aS6DA1CoXIAfxkagYbPjvJgRCTR7NnYQkeK0ebWTjRSS1GyDM2hsSLlKKeM",
-	"Ud/KNDc5w+2vEmzxunqSYpjj8JFI77a/vd/Zq/i6VjTidXFhm5QDUjGa07ZUTS93Op3u25R9TicLi/ej",
-	"PSzssFhkveE4OuxcEhYS1NJfmaLJ0zdH0lII3Gvgme2baWVBX3XqtydStQd+q9d9bGA+LC1r3xWz0gkc",
-	"uZP7EXsjd923UA+Nt9xCYe4OdAXB24Tw2cePPxun2PEwwX8S8z8h1f7isjsL6ADYabaF4OcD71MDMngb",
-	"EofjBWiXuTvC7suQI3QZf8aA7iurBNk/bdpBHklH6ANhzIbyJVUoB70UGcpLpmnBwFdhxQrknaTaF2Qv",
-	"Lz8nCEjqeqCoVG46oLSUErgOW6O+dVudF4WgZlygHIgqJbS2VsHyaKRTXXrdvQaXar3x6VaIzeYaL2ns",
-	"EerLF5oGfa7/OOWQxz9eyusncT3loVlJWnF/4yeHBpKPKJo4skj+dekHXrJIYhvVjyyNuA29XJ7cLRi2",
-	"rULMt8ogrmgxyigVadQwzWAnYsQyvrqTHaZ8B9V3r18aDL7I82hAVPp6LaBoJBpR/OJwt73eFeLhOepd",
-	"0bbMqMrX8ZPLMFT6ct36O6JsB6DQkL1KY7fCwOSh6YxtLWO5OhUiwzBwFDUQLsOO235JRdCsG1/LajWM",
-	"3S4eV816Kc8jOl32t+RaVFuczkx7FmU/n/O2226jvHc6wti+M/8WysyPD8nfwIUZwkcG5LcBjf/F9WeM",
-	"6xO7AzV58A8fNluuyLaXH7boR0HLmk+d1u8qDsdZspO6er0RORqO49HCGXAZvL1/4/abNG9xBkvbdYh0",
-	"u7f9+22J/IAxL6oXMi9i0l69+YxncF8/L65KH/PqBdNgedw9K+88DY2VosWN+n2xUDBQj35Vxej287G9",
-	"ytG1Gl5nQWEPL7Fz5arCYSmZf+eiZpMJKegRHM+PMljhgMND9y9plYVa++922x/tnXlzvflPAAAA//8J",
-	"P7DfZzwAAA==",
+	"jDXw6YHlo51Ww+UiqK2FHe0Bx25IJk1VzojfMf/7SLrdMdYtZSxqq/H69/3WXbTvX8afo3H6V9CBL8/X",
+	"yNYihgP0M9nj6brE3ZA+NoCH5bo3bOZBl5xUF6FBGFQg8BehERj47CgPxkESzZzNEaojhWhzYycaqaUo",
+	"WYbm0NiOcpRTxqhvY5pbnOH2Vwm2cF09RzHMcfhApHfT397r7FV7XRsa8bqwsE3KAakYzWlbqqaPO51O",
+	"923IPqdrhYX7Q/zKIeu/0rkkLCSopb8uRROnb46kpRC418Az2zPTyoK+6tJvT6JqD/xWr/vYcHxYSta+",
+	"J2alEzhyH/cj9jbuOm+hHhpvuYXC3BvoCoJ3CeGTjx9/Nk6x41GC/yTmf0Kq/aVldwbQAbDT7AudDE8N",
+	"yOBdSByOF6Bd1u4Iu69CjtBl/AkDuq+sEmT+tGkFeSQdoQ+EMRvKl1ShHPRSZCgvmaYFA1+BFSuQd5Jq",
+	"X4y9vPycICCp63+iUrnpgNJSSuA6bIv6tm11XhSCmnGBciCqlNDaWgXLo5FOdel19xpcqvW+p1sdNptr",
+	"vKSxR6gvX2Qa9Ln+w5RDHv54Ka+fxPWUh2YlacX9jZ8cGkg+omDiyCL516UfeMkCiW1SP7Is4jb0clfX",
+	"brGwbRVivlUGcQWLUUapSKOGaQY7ESOW8dVd7DDlO6i2e/3SYPAFnkcDotLXawFFI9GIwheHu+21rhAP",
+	"z1HrirZkRlW9jp9chqGyl+vU3xFlq/+FhuxVGrsVBiYPTVdsawnL1agQGYaBo6iBcBl22/ZLKoJG3fg6",
+	"VqtZ7HbxuErWS3ke0emyvyXXntridGbasyj7+Zy33XIb5b3TEcb2Xfm3UGJ+fEj+Bi7MED4yIL8NaPwv",
+	"rj9jXJ/YHajJg3/0sNlyRbZ9/LA9Pwpa1nzqtH5TcTjOkp3U1cuNyNFwHI8WzoDL4N39G7ffpHmHM1ja",
+	"rkOk273t3W9L5AeMeVG9jnkRk/bqzWc8g/v6aXFV+phXr5cGy+PuSXnnWWisFC1u1O+LhYKBevSrKka3",
+	"n47tVY6u1fA6Cwp7eImdK1cVDkvJ/BsXNZtMSEGP4Hh+lMEKBxweun9FqyzU2n+z2/5o78yb681/AgAA",
+	"//9E4hb1YzwAAA==",
 }
 
 // GetSwagger returns the content of the embedded swagger specification file
